refactor: flatten macro crate
