--- crates/macros/impl/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "foundry-macros-impl"
-publish = false
-
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-authors.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lib]
-proc-macro = true
-# proc-macro tests aren't fully supported by cargo-nextest archives
-test = false
-doc = false
-
-[dependencies]
-proc-macro2 = "1.0"
-quote = "1.0"
-syn = "2.0"
-proc-macro-error = "1"
