+++ crates/macros/src/console_fmt.rs
